+++ polodb/__init__.py
@@ -1,7 +1,7 @@
 """Python bindings for the embedded PoloDB document database."""
 
 from ._rust import POLODB_CORE_VERSION, ObjectId, PoloDBError
-from .core import Collection, Cursor, Document, PoloDB, Transaction
+from .core import Collection, Cursor, Document, PoloDB, PoloDBConfig, Transaction
 from .results import DeleteResult, InsertManyResult, InsertOneResult, UpdateResult
 from .version import __version__
 
@@ -15,6 +15,7 @@ __all__ = [
     "InsertOneResult",
     "ObjectId",
     "PoloDB",
+    "PoloDBConfig",
     "PoloDBError",
     "Transaction",
     "UpdateResult",